apps/requirements.txt
requests~=2.31.0
django-ratelimit==4.1.0
djangorestframework~=3.14.0
djangorestframework-simplejwt==5.3.1
Django~=5.0.2
drf-standardized-errors==0.14.0
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
python-amazon-sp-api==1.4.0
django-filter==24.2
backoff==2.2.1
ratelimiter==1.2.0.post0
